requirements.txt
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
sqlparse==0.5.0